+++ anonymize_all_the_things.py
@@ -1,13 +1,9 @@
 #!/usr/bin/env python
 from __future__ import absolute_import, division, print_function, unicode_literals
-from yacryptopan import CryptoPAn
-import netaddr
+from ipaddresscrypto import IPAddressCrypt, printStdErr
 from Crypto import Random #CSPSRNG
 from binascii import hexlify, unhexlify
 import re, sys
-
-def printStdErr(*objs):
-    print(*objs, file=sys.stderr)
 
 printStdErr("generating new random key.")
 key = Random.new().read(32)
@@ -16,23 +12,9 @@ key = Random.new().read(32)
 printStdErr("using key `%s'." % hexlify(key))
 printStdErr("save the key and hard-code it in this file to get reproducible results.")
 
+       
 
-def do_not_anonymize(ip):
-    ip = netaddr.IPAddress(ip)
-    if ip.version == 4:
-        return (ip in netaddr.IPNetwork("10.0.0.0/8") or
-           ip in netaddr.IPNetwork("127.0.0.0/8") or
-           ip in netaddr.IPNetwork("172.16.0.0/12") or
-           ip in netaddr.IPNetwork("192.0.0.0/24") or
-           ip in netaddr.IPNetwork("192.0.2.0/24") or
-           ip in netaddr.IPNetwork("192.168.0.0/16") or
-           ip in netaddr.IPNetwork("224.0.0.0/4"))
-    else:
-        return False #IPv6 addresses can have MACs embedded
-        #be super conservative and anonymize them all
-       
-       
-cp = CryptoPAn(key)
+cp = IPAddressCrypt(key)
 
 def main(filename):
     printStdErr("opening %s" % filename)
@@ -63,19 +45,13 @@ fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|     # fe80::7:8%eth0   fe80::7:8%
         for line in fp:
             for m in ipv6.finditer(line):
                 ip = m.group(0)
-                if not do_not_anonymize(ip):
-                    ip_anonymized = cp.anonymize(ip)
-                    if do_not_anonymize(ip_anonymized):
-                        printStdErr("WARNING: anonymized ip address mapped to special-purpose address range. Please consider re-running with different key")
-                    line = line.replace(ip, ip_anonymized, 1)
+                ip_anonymized = cp.anonymize(ip)
+                line = line.replace(ip, ip_anonymized, 1)
                 
             for m in ipv4.finditer(line):
                 ip = m.group(0)
-                if not do_not_anonymize(ip):
-                    ip_anonymized = cp.anonymize(ip)
-                    if do_not_anonymize(ip_anonymized):
-                        printStdErr("WARNING: anonymized ip address mapped to special-purpose address range. Please consider re-running with different key")
-                    line = line.replace(ip, ip_anonymized, 1)
+                ip_anonymized = cp.anonymize(ip)
+                line = line.replace(ip, ip_anonymized, 1)
             print(line.rstrip('\n'),)
